IL-729 Multi-Cloud uses non-embedded license container
In IL-724 we're moving things to use Instruqt Secrets for license
files. For the Multi-Cloud track switch to using a version of
the `consul-cloud-client` which does not embed a license, and use
the track startup script to read the licenses from Instruqt
Secrets.

This also gets us on the path to "Update the licenses in one place
and everything in Instruqt gets them".